more gatsby-node cleanup + move the refetcher to onCreateDevServer
