--- packages/gatsby-source-wordpress/src/gatsby-node/index.js
@@ -1,13 +1,0 @@
-import onPreBootstrap from "./on-pre-bootstrap"
-import sourceNodes from "./source-nodes"
-import createSchemaCustomization from "./create-schema-customization"
-import createResolvers from "./create-resolvers"
-import onPostBootstrap from "./on-post-bootstrap"
-
-export default {
-  onPreBootstrap,
-  sourceNodes,
-  createSchemaCustomization,
-  createResolvers,
-  onPostBootstrap,
-}
