+++ packages/gatsby-source-wordpress/src/gatsby-node/on-create-dev-server.js
